+++ Editor/AnimationRig/Constraints/MultiPositionConstraintEditor.cs
@@ -56,7 +56,15 @@ namespace UnityEditor.Animations.Rigging
 
             m_ReorderableList.onChangedCallback = (ReorderableList reorderableList) =>
             {
+                Undo.RegisterCompleteObjectUndo(m_Constraint, "Edit MultiPosition");
                 m_Constraint.data.sourceObjects = (WeightedTransformArray)reorderableList.list;
+                if (PrefabUtility.IsPartOfPrefabInstance(m_Constraint))
+                    EditorUtility.SetDirty(m_Constraint);
+            };
+
+            Undo.undoRedoPerformed += () =>
+            {
+                m_ReorderableList.list = m_Constraint.data.sourceObjects;
             };
         }